Remove linter rule for duplicate key checks, move back to validation
